# api/openapi.yaml
openapi: 3.0.3
info:
  title: Barcode BFF API
  version: 1.0.0
  description: |
    Актуальная OpenAPI-спецификация для HTTP API BFF.

    BFF выступает оркестратором и не содержит доменной логики расчётов.
    Все non-system endpoints могут дополнительно вернуть 503 при включённом
    maintenance mode.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /api/v1/upload:
    post:
      tags: [Bulk]
      summary: Загрузка XLS-файла для массовой генерации (ТЗ §3.1)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: XLS-файл с данными для генерации
                revision:
                  type: string
                  description: Ревизия шаблона (по умолчанию US_CA_08292017)
      responses:
        '200':
          description: Файл принят, batch создан и готов к подтверждению
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
              example:
                success: true
                batchId: batch-123
                status: ready
                summary:
                  totalRows: 100
                  validRows: 95
                  invalidRows: 5
                billing:
                  canProcess: true
                  requested: 95
                  allowedTotal: 95
                nextStep: confirm
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '402':
          description: Недостаточно средств — частичная обработка (PARTIAL_AVAILABLE)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: PARTIAL_AVAILABLE
                message: "Insufficient funds. Generate 60 of 100?"
                details:
                  batchId: batch-123
                  billing:
                    canProcess: true
                    requested: 100
                    allowedTotal: 60
                  options:
                    generatePartial:
                      count: 60
                    cancel:
                      message: Cancel batch and top up balance
        '503':
          $ref: '#/components/responses/ServiceError'
  /api/v1/batches:
    get:
      tags: [Bulk]
      summary: Список батчей текущего пользователя (ТЗ §3.6)
      parameters:
        - in: query
          name: status
          required: false
          schema:
            type: string
            enum: [pending, ready, processing, completed, cancelled, failed]
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
          description: RFC3339 начало диапазона
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
          description: RFC3339 конец диапазона
        - in: query
          name: revision
          required: false
          schema:
            type: string
        - in: query
          name: sortBy
          required: false
          schema:
            type: string
        - in: query
          name: order
          required: false
          schema:
            type: string
            enum: [asc, desc]
        - in: query
          name: cursor
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Список батчей
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchListResponse'
              example:
                data:
                  - batchId: batch-123
                    status: completed
                    totalJobs: 100
                    completedJobs: 98
                    failedJobs: 2
                    createdAt: '2026-01-08T10:00:00Z'
                    completedAt: '2026-01-08T10:15:00Z'
                meta:
                  page: 1
                  limit: 20
                  total: 1
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/batch/{batchId}:
    get:
      tags: [Bulk]
      summary: Статус и прогресс батча (ТЗ §3.4)
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Состояние батча
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/RequestInFlightError'
  /api/v1/batch/{batchId}/status:
    get:
      tags: [Bulk]
      summary: Алиас GET /api/v1/batch/{batchId} — статус батча (ТЗ §3.4)
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Состояние батча
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchStatusResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          $ref: '#/components/responses/RequestInFlightError'
  /api/v1/batch/{batchId}/confirm:
    post:
      tags: [Bulk]
      summary: Подтверждение генерации батча (ТЗ §3.3)
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConfirmRequest'
            example:
              action: generate_all
      responses:
        '202':
          description: Батч принят в обработку
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConfirmResponse'
              example:
                success: true
                batchId: batch-123
                status: processing
                jobsQueued: 95
                estimatedTimeSeconds: 285
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '402':
          $ref: '#/components/responses/InsufficientFundsError'
  /api/v1/batch/{batchId}/cancel:
    post:
      tags: [Bulk]
      summary: Отмена батча (ТЗ §3.5)
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Батч отменён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResponse'
              example:
                success: true
                batchId: batch-123
                status: cancelled
                refund:
                  pending: 50
                  refunded: 50
                  transactionIds:
                    - tx-1
                    - tx-2
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '503':
          $ref: '#/components/responses/ServiceError'
  /api/v1/batch/{batchId}/export:
    post:
      tags: [Bulk]
      summary: Запустить асинхронный экспорт результатов батча
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        '202':
          description: Экспорт поставлен в очередь
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportStartResponse'
              example:
                id: batch-123
                status: queued
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /api/v1/export/{id}:
    get:
      tags: [Bulk]
      summary: Получить статус/ссылку на готовый экспорт
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: query
          name: download
          required: false
          schema:
            type: boolean
          description: Если true — проксирует файл вместо presigned URL
      responses:
        '200':
          description: Статус экспорта, presigned URL или бинарный файл
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExportGetResponse'
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/batch/{batchId}/errors.xlsx:
    get:
      tags: [Bulk]
      summary: Скачать файл ошибок валидации (ТЗ §3.7)
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: XLSX-файл с ошибками валидации
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/batch/{batchId}/results.xlsx:
    get:
      tags: [Bulk]
      summary: Скачать результаты батча в XLSX (ТЗ §3.8)
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: XLSX-файл с результатами (buildId, barcodeUrls)
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/template/{revision}:
    get:
      tags: [Bulk]
      summary: Скачать XLS-шаблон для ревизии (ТЗ §5.4)
      security: []
      parameters:
        - in: path
          name: revision
          required: true
          schema:
            type: string
          description: Идентификатор ревизии, например US_CA_08292017
      responses:
        '200':
          description: XLS-шаблон
          content:
            application/vnd.ms-excel:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/batches:
    get:
      tags: [Admin]
      summary: Список всех батчей (ТЗ §10.5.1)
      security:
        - adminToken: []
      parameters:
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: perPage
          required: false
          schema:
            type: integer
            minimum: 1
            default: 20
        - in: query
          name: status
          required: false
          schema:
            type: string
        - in: query
          name: userId
          required: false
          schema:
            type: string
        - in: query
          name: revision
          required: false
          schema:
            type: string
        - in: query
          name: from
          required: false
          schema:
            type: string
            format: date-time
        - in: query
          name: to
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Список батчей с пагинацией
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBatchListResponse'
              example:
                batches:
                  - id: batch-123
                    userId: user-456
                    status: processing
                    totalJobs: 100
                    completedJobs: 45
                    failedJobs: 2
                    createdAt: '2026-01-08T10:00:00Z'
                pagination:
                  page: 1
                  perPage: 20
                  total: 150
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /api/v1/admin/batches/{batchId}:
    get:
      tags: [Admin]
      summary: Детали батча (ТЗ §10.5.3)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Детальная информация по батчу включая jobs
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminBatchDetailResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/batches/{batchId}/config:
    put:
      tags: [Admin]
      summary: Обновить конфигурацию батча (ТЗ §10.5.5)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminUpdateBatchConfigRequest'
            example:
              priority: high
              timeout: 30000
      responses:
        '200':
          description: Конфигурация батча обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminUpdateBatchConfigResponse'
              example:
                success: true
                batchId: batch-123
                config:
                  priority: high
                  timeout: 30000
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /api/v1/admin/batches/{batchId}/restart:
    post:
      tags: [Admin]
      summary: Перезапустить упавшие jobs батча (ТЗ §10.5.4)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: batchId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RestartRequest'
            example:
              restartMode: failed_only
              force: false
      responses:
        '200':
          description: Результат перезапуска
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RestartResponse'
              example:
                restarted: 12
                skipped: 3
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /api/v1/admin/stats:
    get:
      tags: [Admin]
      summary: Получить административную статистику (агрегаты за период from..to)
      security:
        - adminToken: []
      parameters:
        - in: query
          name: from
          required: false
          description: RFC3339 начало периода (по умолчанию - начало дня UTC)
          schema:
            type: string
        - in: query
          name: to
          required: false
          description: RFC3339 конец периода (по умолчанию - now UTC)
          schema:
            type: string
      responses:
        '200':
          description: Admin stats
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminStatsResponse'
              example:
                totalBatches: 1500
                activeBatches: 15
                completedToday: 45
                failedToday: 3
                avgProcessingTime: 125000
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '503':
          $ref: '#/components/responses/ServiceError'
        '409':
          $ref: '#/components/responses/RequestInFlightError'
...
  /api/v1/admin/config/bulk-limits:
    get:
      tags: [Admin]
      summary: Получить конфиг лимитов для bulk
      security:
        - adminToken: []
      responses:
        '200':
          description: Bulk limits config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BulkLimitsConfig'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    put:
      tags: [Admin]
      summary: Обновить конфиг лимитов для bulk
      security:
        - adminToken: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BulkLimitsConfig'
      responses:
        '200':
          description: Updated bulk limits config; повтор с тем же `X-Idempotency-Key` возвращает cached response с `code=DUPLICATE_REQUEST`
          content:
            application/json:
              examples:
                success:
                  value:
                    enabled: true
                    defaultPerUser: 10
                    maxPerBatch: 1000
                    perRevision:
                      US_CA_08292017: 20
              schema:
                $ref: '#/components/schemas/BulkLimitsConfig'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '409':
          $ref: '#/components/responses/RequestInFlightError'
...
  /api/v1/admin/users/{userId}/batches:
    get:
      tags: [Admin]
      summary: Список батчей для указанного пользователя (админ)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: query
          name: page
          required: false
          schema:
            type: integer
            minimum: 1
        - in: query
          name: perPage
          required: false
          schema:
            type: integer
            minimum: 1
        - in: query
          name: status
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Batches for user
          content:
            application/json:
              examples:
                sample:
                  summary: Пример ответа с одним батчем
                  value:
                    batches:
                      - id: batch-123
                        userId: user-456
                        status: processing
                        totalJobs: 100
                        completedJobs: 45
                        failedJobs: 2
                        createdAt: '2026-01-08T10:00:00Z'
                    pagination:
                      page: 1
                      perPage: 20
                      total: 1
              schema:
                $ref: '#/components/schemas/UserBatchesResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    adminToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InsufficientFundsError:
      description: Insufficient funds or free edit already used
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceError:
      description: Downstream service error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Missing or invalid bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SimpleErrorResponse'
    ForbiddenError:
      description: Invalid admin or service token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/SimpleErrorResponse'
    RequestInFlightError:
      description: A request with the same idempotency key is already being processed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required: [code, message]
      example:
        code: VALIDATION_ERROR
        message: "required input fields missing: [lastName]"
        details:
          missingFields:
            - lastName
      properties:
        code:
          type: string
          enum:
            - VALIDATION_ERROR
            - MISSING_DEPENDENCY
            - INSUFFICIENT_FUNDS
            - PARTIAL_FUNDS
            - DUPLICATE_REQUEST
            - BARCODEGEN_ERROR
            - BILLING_ERROR
            - REQUEST_IN_FLIGHT
            - PARTIAL_AVAILABLE
            - SERVICE_ERROR
            - BATCH_NOT_FOUND
            - BATCH_ALREADY_PROCESSED
            - BATCH_CANCELLED
            - NOT_FOUND
        message:
          type: string
        details:
          type: object
          additionalProperties: true
...
    UploadResponse:
      type: object
      required: [success, batchId, status]
      properties:
        success:
          type: boolean
        batchId:
          type: string
        status:
          type: string
          enum: [ready, partial_available]
        summary:
          $ref: '#/components/schemas/UploadSummary'
        billing:
          type: object
          additionalProperties: true
        nextStep:
          type: string
          example: confirm
...
    BatchListResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/BatchSummary'
        meta:
          type: object
          required: [page, limit, total]
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
...
    BatchStatusResponse:
      type: object
      required: [batchId, status]
      example:
        batchId: batch-123
        status: processing
        progress:
          total: 100
          completed: 45
          failed: 2
          pending: 53
          percentComplete: 45.0
        billing:
          charged: 45
          refunded: 2
          pending: 53
        createdAt: '2026-01-08T10:00:00Z'
        updatedAt: '2026-01-08T10:15:00Z'
      properties:
        batchId:
          type: string
        status:
          type: string
        progress:
          type: object
          properties:
            total:
              type: integer
            completed:
              type: integer
            failed:
              type: integer
            pending:
              type: integer
            percentComplete:
              type: number
              format: double
        results:
          type: object
          properties:
            successful:
              type: array
              items:
                type: object
                properties:
                  row:
                    type: integer
                  buildId:
                    type: string
                  barcodeUrls:
                    type: object
                    properties:
                      pdf417:
                        type: string
                      code128:
                        type: string
            failed:
              type: array
              items:
                type: object
                properties:
                  row:
                    type: integer
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
        billing:
          type: object
          properties:
            charged:
              type: integer
            refunded:
              type: integer
            pending:
              type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ConfirmRequest:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [generate_all, generate_partial, generate_valid, cancel]
        count:
          type: integer
          minimum: 1
          description: Обязателен при action=generate_partial
    ConfirmResponse:
      type: object
      required: [success, batchId, status, jobsQueued, estimatedTimeSeconds]
      properties:
        success:
          type: boolean
        batchId:
          type: string
        status:
          type: string
          example: processing
        jobsQueued:
          type: integer
        estimatedTimeSeconds:
          type: integer
    CancelResponse:
      type: object
      required: [success, batchId, status, refund]
      properties:
        success:
          type: boolean
        batchId:
          type: string
        status:
          type: string
          example: cancelled
        refund:
          type: object
          required: [pending, refunded, transactionIds]
          properties:
            pending:
              type: integer
            refunded:
              type: integer
            transactionIds:
              type: array
              items:
                type: string
    ExportStartResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [queued, processing, done, failed]
    ExportGetResponse:
      type: object
      required: [id, status]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [queued, processing, done, failed]
        url:
          type: string
          description: Presigned URL для скачивания
        expires_at:
          type: string
          format: date-time
        storage_id:
          type: string
        error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AdminBatchListResponse:
      type: object
      required: [batches, pagination]
      properties:
        batches:
          type: array
          items:
            $ref: '#/components/schemas/AdminBatchSummary'
        pagination:
          type: object
          required: [page, perPage, total]
          properties:
            page:
              type: integer
            perPage:
              type: integer
            total:
              type: integer
...
    AdminUpdateBatchConfigRequest:
      type: object
      properties:
        priority:
          type: string
          description: Приоритет обработки (например high, normal, low)
        timeout:
          type: integer
          description: Таймаут в миллисекундах
    AdminUpdateBatchConfigResponse:
      type: object
      required: [success, batchId, config]
      properties:
        success:
          type: boolean
        batchId:
          type: string
        config:
          type: object
          required: [priority, timeout]
          properties:
            priority:
              type: string
            timeout:
              type: integer
...
    RestartResponse:
      type: object
      required: [restarted, skipped]
      properties:
        restarted:
          type: integer
          description: Количество перезапущенных jobs
        skipped:
          type: integer
          description: Количество пропущенных jobs